+++ Payee/CurrencyManager.m
@@ -20,7 +20,7 @@
 
 @implementation CurrencyManager
 
-+(double)convert: (float)foreignValue
++(float)convert: (float)foreignValue
             from: (NSString*)foreignCurrency
               to: (NSString*)baseCurrency
 {
